+++ cmd/mdatagen/metricdata_test.go
@@ -30,7 +30,7 @@ func TestMetricData(t *testing.T) {
 		{&intGauge{}, "IntGauge", false, false},
 		{&intSum{}, "IntSum", true, true},
 		{&intHistogram{}, "IntHistogram", true, false},
-		{&doubleGauge{}, "DoubleGauge", false, false},
+		{&gauge{}, "Gauge", false, false},
 		{&sum{}, "Sum", true, true},
 		{&histogram{}, "Histogram", true, false},
 	} {