+++ internal/goldendataset/pict_metrics_gen.go
@@ -73,7 +73,7 @@ func pictToCfg(inputs PICTMetricInputs) MetricsCfg {
 		cfg.MetricDescriptorType = pdata.MetricDataTypeIntSum
 		cfg.IsMonotonicSum = false
 	case MetricTypeDoubleGauge:
-		cfg.MetricDescriptorType = pdata.MetricDataTypeDoubleGauge
+		cfg.MetricDescriptorType = pdata.MetricDataTypeGauge
 	case MetricTypeMonotonicDoubleSum:
 		cfg.MetricDescriptorType = pdata.MetricDataTypeSum
 		cfg.IsMonotonicSum = true