+++ model/pdata/generated_metrics.go
@@ -589,35 +589,35 @@ func (ms IntGauge) CopyTo(dest IntGauge) {
 	ms.DataPoints().CopyTo(dest.DataPoints())
 }
 
-// DoubleGauge represents the type of a double scalar metric that always exports the "current value" for every data point.
+// Gauge represents the type of a double scalar metric that always exports the "current value" for every data point.
 //
 // This is a reference type, if passed by value and callee modifies it the
 // caller will see the modification.
 //
-// Must use NewDoubleGauge function to create new instances.
+// Must use NewGauge function to create new instances.
 // Important: zero-initialized instance is not valid for use.
-type DoubleGauge struct {
+type Gauge struct {
 	orig *otlpmetrics.Gauge
 }
 
-func newDoubleGauge(orig *otlpmetrics.Gauge) DoubleGauge {
-	return DoubleGauge{orig: orig}
+func newGauge(orig *otlpmetrics.Gauge) Gauge {
+	return Gauge{orig: orig}
 }
 
-// NewDoubleGauge creates a new empty DoubleGauge.
+// NewGauge creates a new empty Gauge.
 //
 // This must be used only in testing code since no "Set" method available.
-func NewDoubleGauge() DoubleGauge {
-	return newDoubleGauge(&otlpmetrics.Gauge{})
+func NewGauge() Gauge {
+	return newGauge(&otlpmetrics.Gauge{})
 }
 
-// DataPoints returns the DataPoints associated with this DoubleGauge.
-func (ms DoubleGauge) DataPoints() DoubleDataPointSlice {
+// DataPoints returns the DataPoints associated with this Gauge.
+func (ms Gauge) DataPoints() DoubleDataPointSlice {
 	return newDoubleDataPointSlice(&(*ms.orig).DataPoints)
 }
 
 // CopyTo copies all properties from the current struct to the dest.
-func (ms DoubleGauge) CopyTo(dest DoubleGauge) {
+func (ms Gauge) CopyTo(dest Gauge) {
 	ms.DataPoints().CopyTo(dest.DataPoints())
 }