+++ receiver/prometheusreceiver/internal/otlp_metricsbuilder_test.go
@@ -84,7 +84,7 @@ func TestGetBoundaryEquivalence(t *testing.T) {
 		{
 			name:   "other data types without matches",
 			mtype:  metricspb.MetricDescriptor_GAUGE_DOUBLE,
-			pmtype: pdata.MetricDataTypeDoubleGauge,
+			pmtype: pdata.MetricDataTypeGauge,
 			labels: labels.Labels{
 				{Name: model.BucketLabel, Value: "11.71"},
 			},
@@ -161,7 +161,7 @@ func TestGetBoundaryPdata(t *testing.T) {
 		},
 		{
 			name:  "other data types without matches",
-			mtype: pdata.MetricDataTypeDoubleGauge,
+			mtype: pdata.MetricDataTypeGauge,
 			labels: labels.Labels{
 				{Name: model.BucketLabel, Value: "11.71"},
 			},
@@ -204,7 +204,7 @@ func TestConvToPdataMetricType(t *testing.T) {
 		{
 			name:  "textparse.unknown",
 			mtype: textparse.MetricTypeUnknown,
-			want:  pdata.MetricDataTypeDoubleGauge,
+			want:  pdata.MetricDataTypeGauge,
 		},
 		{
 			name:  "textparse.histogram",
@@ -251,7 +251,7 @@ func TestIsusefulLabelPdata(t *testing.T) {
 			},
 			mtypes: []pdata.MetricDataType{
 				pdata.MetricDataTypeSum,
-				pdata.MetricDataTypeDoubleGauge,
+				pdata.MetricDataTypeGauge,
 				pdata.MetricDataTypeIntHistogram,
 				pdata.MetricDataTypeHistogram,
 				pdata.MetricDataTypeSummary,
@@ -272,7 +272,7 @@ func TestIsusefulLabelPdata(t *testing.T) {
 			name: `bucket label with non "int_histogram", "histogram":: useful`,
 			mtypes: []pdata.MetricDataType{
 				pdata.MetricDataTypeSum,
-				pdata.MetricDataTypeDoubleGauge,
+				pdata.MetricDataTypeGauge,
 				pdata.MetricDataTypeSummary,
 				pdata.MetricDataTypeIntSum,
 				pdata.MetricDataTypeNone,
@@ -295,7 +295,7 @@ func TestIsusefulLabelPdata(t *testing.T) {
 			labelKeys: []string{model.QuantileLabel},
 			mtypes: []pdata.MetricDataType{
 				pdata.MetricDataTypeSum,
-				pdata.MetricDataTypeDoubleGauge,
+				pdata.MetricDataTypeGauge,
 				pdata.MetricDataTypeIntHistogram,
 				pdata.MetricDataTypeHistogram,
 				pdata.MetricDataTypeIntSum,
@@ -310,7 +310,7 @@ func TestIsusefulLabelPdata(t *testing.T) {
 			labelKeys: []string{"any_label", "foo.bar"},
 			mtypes: []pdata.MetricDataType{
 				pdata.MetricDataTypeSum,
-				pdata.MetricDataTypeDoubleGauge,
+				pdata.MetricDataTypeGauge,
 				pdata.MetricDataTypeIntHistogram,
 				pdata.MetricDataTypeHistogram,
 				pdata.MetricDataTypeSummary,