+++ testutil/metricstestutil/metricsutil.go
@@ -172,9 +172,9 @@ func SortedMetrics(metrics pdata.Metrics) pdata.Metrics {
 					for l := 0; l < m.IntSum().DataPoints().Len(); l++ {
 						m.IntSum().DataPoints().At(l).LabelsMap().Sort()
 					}
-				case pdata.MetricDataTypeDoubleGauge:
-					for l := 0; l < m.DoubleGauge().DataPoints().Len(); l++ {
-						m.DoubleGauge().DataPoints().At(l).LabelsMap().Sort()
+				case pdata.MetricDataTypeGauge:
+					for l := 0; l < m.Gauge().DataPoints().Len(); l++ {
+						m.Gauge().DataPoints().At(l).LabelsMap().Sort()
 					}
 				case pdata.MetricDataTypeSum:
 					for l := 0; l < m.Sum().DataPoints().Len(); l++ {